backend/requirements.txt
alembic==1.5.8
aniso8601==9.0.1
astroid==2.5.6
click==7.1.2
colorama==0.4.4
coverage==5.5
Flask==1.1.2
Flask-Caching==1.10.1
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.3
mccabe==0.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==2.1.0
pylint==2.8.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
six==1.15.0
SQLAlchemy==1.4.11
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
